+++ nall/varint.hpp
@@ -5,18 +5,18 @@
 #include <nall/static.hpp>
 
 namespace nall {
-  template<unsigned bits> class uint_t {
+  template<uint64_t bits> class uint_t {
   private:
     enum { bytes = (bits + 7) >> 3 };  //minimum number of bytes needed to store value
     typedef typename static_if<
       sizeof(int) >= bytes,
-      unsigned int,
+      uint64_t,
       typename static_if<
         sizeof(long) >= bytes,
-        unsigned long,
+        uint64_t long,
         typename static_if<
           sizeof(long long) >= bytes,
-          unsigned long long,
+          uint64_t long long,
           void
         >::type
       >::type
@@ -45,7 +45,7 @@ namespace nall {
     inline uint_t(const T i) : data(uclip<bits>(i)) {}
   };
 
-  template<unsigned bits> class int_t {
+  template<uint64_t bits> class int_t {
   private:
     enum { bytes = (bits + 7) >> 3 };  //minimum number of bytes needed to store value
     typedef typename static_if<
